repo: organize Python modules into src directory (#139)
