+++ src/assets.py
